Rest api integration (#46)
* Implemented Users Access Tokens

Now Users can generate their access tokens in a dedicated page. It is still a draft and needs to be tested, but it's a good starting point.

* UrlController Short URL validation

Edited UrlController Short URL validation so now we use an extended validation and we have more flexibility, allowing us to use it for APIs too.

* First APIs

* Url Management APIs

All URL management APIs created. They need to be so much improved, though.

* Apply fixes from StyleCI (#45)

* Update UrlController.php